Email templates for moderation
Fixes #66

* Moderation email honours microsite of pledge

Some circumlocution, due to microsite not being set within the admin.

* Localise email site name and sender

Note that the Admin page doesn't tend to have a microsite set: you can
explicitly do this, but to avoid missing anything, in single-customer
sites (such as the RBWM case) it may be easier to use the new
OPTION_FIXED_SITE_NAME override.