RBWM hide password change box until button is clicked
as per dave's suggestion, hide the box only for Javascript aware browsers.

Am doing this by emitting style using document.write
(I prefer this to onload { box.hide() } as there isn't an
awkward jump.

Have also added code to move the success/failure message out of the
display:none box (there is a jump in this case, but as it's for
an information message, this could be considered a feature?)